feat: add `std.path.filter`
Easily filter directories based on a simple whitelist of paths:
```
std.path.filter [
  "some/dir"
  "some/other/file"
] ./some/dir
```

`std.set.filter` returns a store path with the filtered directories
contents. This is useful for quickly collecting paths in a large
repository, e.g. a mono-repo.

If the second argument is not a directory, it will simply return a store
path with an empty file (same as passing a constant `false` value to a
`builtins.path` filter.

If a directory is in the list, it's entire contents are added.

This function is used to offer a simple API in the TOML manifest to
include filtered source files as dependencies available at `atom.src`:
```toml
[src]
binSrcA = ["paths" "to" "include"]
binSrcB = ["some" "other" "paths"]
```